requirements.txt
Django==3.2.9
django-ckeditor==6.2.0
Pillow==9.0.0